+++ common-components/utils/multiDomain-utils.ts
@@ -1,0 +1,8 @@
+import useStore from "../store";
+
+export const getDomainsArray = (currentDomains) => {
+    const multiDomainEnabled = import.meta.env.REACT_APP_ENABLE_MULTI_DOMAIN.toLowerCase() === 'true';
+    const userDomains = currentDomains || [];
+
+    return multiDomainEnabled ? (userDomains?.length > 0 ? userDomains : [null]) : [];
+}